+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # 0.1.0
 
 Initial Version of the library.
+
+# 0.2.0
+
+Change getTransformer to more universal fromSubtree, now it should be used instead getState and getActions
+Change putTransformer to more universal putIntoSubtree now it should be used instead putState and putActions